Protocol exception ':status' header not present
Retrofit network calls fails with a Protocol Exception suddenly in a working app.
The app was working till yesterday and today all the network calls fails.
The calls works fine with HTTP but fails with HTTPS.

After hours of mess, finally got a solution. Updating Retrofit libraries to the
latest version did the trick for me.

Signed-off-by: Solo <anqi.huang@outlook.com>